cmd/snap, cmd/snapd, daemon/snapd: make snapd a symlink to snap
Crude attempt at reducing the size of binaries occupied by the snapd package.
The hack makes the snapd binary a symlink to /usr/bin/snap.

Before we had snapd (24MB) and snap (20MB) binaries. With the change we have a
single binary snap (26MB). Around 18MB is saved.
